Add booking back to pages
